Added an activity to view a channel.
The idea is that RSS channel stores more information than we can show
in the main feed list.  This activity lets a user interact with a specific
channel and see things like the description, image, and link.  It  also lists
the images items and should allow the user to interact with those items
the same way the do in the main tree view based activity.