Changed date in test data.
Changed the date of the example wait-wait example podcast from 
Jan 01, 2011 to Jan 02, 2011.  Jan 01 has the same numberic value,
which means testing for a date of 2011-01-01 and passing is less of a
test than passing for 2011-01-02 and passing. In the former case, the
backend could reverse the month and day and still pass.