Changed update to not nuke previous entries.
The previous version deleted all the item entries from the
Item db table and just reloaded everything.  This version
checks to see if the item is in the table and only adds
the new items.

The immediate advantage of this is that doing an update without
a connection will NOT clear out the Item table, which it was before.