Multiple code cleanups.
Added a static method create item view intents.

Moved creation of the intent to view a single item to class
that will handle the intent.  This encapsulates the logic
better than just creating the intent whereever we want to dispatch it.

Removed an irrelevant todo

Made string array init consistent with other string array init.

Removed an unneed import on Castroid.

Import was to get the TAG field in Castroid.
Used the TAG field in the dataprovider instead.

Removed old classes.

Deleted classes were from initial testing and are no longer in use.