Fixed bug in add feed.
Podcast would not be added if a url was 'checked' before being saved.
Incorrect state check in saveRSSFeed reported that a check was in
progress if a feed had been checked and check was finished.

Update corrects this bug and changes the save feed so that it does not
reparse the feed, as long as the url entered has not changed.