Fixed finding main RSS element in android > 2.3.
The struture that the RSS document is parsed into must be represented
differently after 2.3.

Previously, the document element from processing an XML file was an
element that contained the rss node. After 2.3 the document element is
the rss.

Change checks to see if the node, n, is the rss elemeent and just returns
n in that case. Otherwise, the algorithm procedes as before,
looking for the rss node as child of n.